requirements.txt
# PWA & IWT Wave Tour Stats Scraper Dependencies

# Web Scraping
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Date/Time Utilities
python-dateutil>=2.8.0

# WebDriver Management
webdriver-manager>=4.0.0

# Environment Variables & Database
python-dotenv>=1.0.0
mysql-connector-python>=8.3.0
pyodbc>=5.0.0  # Legacy - not actively used
